requirements.txt
streamlit
pymongo
bcrypt
python-docx
google-generativeai